requirements.txt
Django==4.2.7
Pillow==10.4.0
python-decouple==3.8
django-crispy-forms==2.0
crispy-bootstrap5==0.7
django-extensions==3.2.3
django-debug-toolbar==4.2.0
django-environ==0.11.2
djangorestframework==3.14.0
django-cors-headers==4.3.1

